pyproject.toml
[tool.poetry]
name = "insighter-telegram-bot"
version = "0.1.0"
description = ""
authors = ["artemprod <artemprod1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nats-py = "^2.7.2"
pydantic = "^2.7.1"
aiohttp = "^3.9.5"
aiogram = "^3.5.0"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
aiobotocore = "^2.13.0"
requests = "^2.32.2"
types-aiobotocore = {extras = ["aiobotocore"], version = "^2.13.0"}
aioselectel-api = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
